Fix docs style
