+++ import_map.json
@@ -2,6 +2,7 @@
   "imports": {
     "components/": "./components/",
     "std/": "https://deno.land/std@0.140.0/",
+    "npm/": "https://esm.sh/",
     "@unocss/": "https://deno.land/x/aleph@1.0.0-beta.2/lib/@unocss/",
     "aleph/": "https://deno.land/x/aleph@1.0.0-beta.2/",
     "aleph/react": "https://deno.land/x/aleph@1.0.0-beta.2/runtime/react/mod.ts",
