+++ index.html
@@ -5,6 +5,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link rel="icon" href="./assets/logo.svg">
+  <link href="/gfm.css" rel="stylesheet">
 </head>
 
 <body>
