+++ routes/index.tsx
@@ -9,10 +9,10 @@ const keywords = [
   "alephjs",
   "aleph.js",
   "react",
+  "vue",
   "full-stack",
   "framework",
   "ssr",
-  "ssg",
   "deno",
   "typescript",
   "out-of-the-box",
