+++ utils/gfm.ts
@@ -1,0 +1,8 @@
+import { CSS, render } from "https://deno.land/x/gfm@0.1.20/mod.ts";
+import { safeLoadFront } from "npm/yaml-front-matter@4.1.1";
+
+import "https://esm.sh/prismjs@1.27.0/components/prism-jsx?no-check";
+import "https://esm.sh/prismjs@1.27.0/components/prism-bash?no-check";
+import "https://esm.sh/prismjs@1.27.0/components/prism-typescript?no-check";
+
+export { CSS, render, safeLoadFront };
